Fix some bugs
resortColumns needs to update the new (in v4.0.0) jqgfirstrow even
when we're going to empty and refill the grid, because its used as
a template for new rows in the table.

In jquery 3.5, you can no longer do "<span/>" and expect it to work.
It needs to be "<span></span>".